PythonProject5/requirements.txt
fastapi>=0.115.8
uvicorn[standard]>=0.34.0
python-dotenv==1.0.0
openai>=1.61.1
llama-index==0.5.6
llama-index-core>=0.12.16.post1
qdrant-client>=1.13.2
pymongo>=4.11
PyMuPDF>=1.25.3
python-docx>=1.1.2
llama-parse>=0.5.20